feat: 체력단련 버튼 제작 완료
1. 체력단련 버튼, 레벨업 버튼 및 기능 제작 완료.
2. DB 숫자들로만 보이게 설정 완료
3. DB정보들 중 값이 큰것들은 double로 변환 완료

+++ Cat_Merge/Assets/1.Scripts/GameManagement/MouseDataLoader.cs
@@ -60,8 +60,8 @@ public class MouseDataLoader : MonoBehaviour
                 int mouseId = int.Parse(validValues[0]);
                 string mouseName = validValues[1];
                 int mouseGrade = int.Parse(validValues[2]);
-                int mouseDamage = int.Parse(validValues[3]);
-                float mouseHp = float.Parse(validValues[4]);
+                double mouseDamage = double.Parse(validValues[3]);
+                double mouseHp = double.Parse(validValues[4]);
                 Sprite mouseImage = LoadSprite(validValues[5]);
                 int numOfAttack = int.Parse(validValues[6]);
                 int mouseAttackSpeed = int.Parse(validValues[7]);
